feat: Update data models and DTOs for improved API integration and error handling

+++ backend/src/main/java/ticketmaster/proyecto/repository/TipoEstablecimientosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ticketmaster.proyecto.model.TipoEstablecimiento;
 
-public interface TipoEstablecimientosRepository extends JpaRepository<TipoEstablecimiento,Integer> {
+public interface TipoEstablecimientosRepository extends JpaRepository<TipoEstablecimiento,Long> {
 
 }
